app/requirements.txt
tensorflow
fastapi
numpy
pillow
uvicorn
